Refactor to single package
