+++ apps/api/package.json
@@ -26,8 +26,7 @@
 	},
 	"dependencies": {
 		"@pcgl-daco/data-model": "workspace:^",
-		"@pcgl-daco/logger-express-middleware": "workspace:^",
-		"@pcgl-daco/logger-client": "workspace:^",
+		"@pcgl-daco/logger": "workspace:^",
 		"@types/body-parser": "^1.19.5",
 		"body-parser": "^1.20.3",
 		"cors": "^2.8.5",
