--- packages/logger/client/package.json
@@ -1,25 +1,0 @@
-{
-	"name": "@pcgl-daco/logger-client",
-	"version": "0.1.0",
-	"description": "OHCRN Platform Logger",
-	"exports": "./dist/index.js",
-	"types": "dist/index.d.ts",
-	"type": "module",
-	"engines": {
-		"node": ">=18"
-	},
-	"scripts": {
-		"build": "tsc"
-	},
-	"keywords": [],
-	"author": "",
-	"license": "ISC",
-	"dependencies": {
-		"winston": "^3.10.0"
-	},
-	"devDependencies": {
-		"dotenv": "^16.3.1",
-		"@pcgl-daco/logger-eslint-config": "workspace:^",
-		"@pcgl-daco/logger-types": "workspace:^"
-	}
-}
