--- packages/logger/client/tsconfig.json
@@ -1,10 +1,0 @@
-{
-	"extends": "../../../tsconfig.base.json",
-	"compilerOptions": {
-		"outDir": "dist",
-		"rootDir": "src",
-		"baseUrl": ".",
-		"declaration": true
-	},
-	"include": ["src"]
-}
