--- packages/logger/eslint-config/package.json
@@ -1,14 +1,0 @@
-{
-	"name": "@pcgl-daco/logger-eslint-config",
-	"version": "1.0.0",
-	"main": "index.js",
-	"devDependencies": {
-		"@typescript-eslint/eslint-plugin": "^6.2.0",
-		"@typescript-eslint/parser": "^6.2.0",
-		"eslint": "8.40.0",
-		"eslint-config-prettier": "^8.8.0",
-		"eslint-import-resolver-typescript": "^3.6.1",
-		"eslint-plugin-import": "^2.29.1",
-		"eslint-plugin-prettier": "^5.0.0"
-	}
-}
