--- packages/logger/express-middleware/.eslintrc.json
@@ -1,3 +1,0 @@
-{
-	"extends": ["ohcrn"]
-}
