+++ packages/logger/package.json
@@ -1,0 +1,38 @@
+{
+	"name": "@pcgl-daco/logger",
+	"version": "0.1.0",
+	"description": "PCGL-DACO Logger",
+	"exports": "./dist/index.js",
+	"types": "dist/index.d.ts",
+	"type": "module",
+	"engines": {
+		"node": ">=18"
+	},
+	"scripts": {
+		"build": "tsc"
+	},
+	"keywords": [],
+	"author": "",
+	"license": "ISC",
+	"dependencies": {
+		"winston": "^3.10.0",
+		"express-requests-logger": "^4.0.0"
+	},
+	"devDependencies": {
+		"dotenv": "^16.3.1",
+		"@anatine/zod-openapi": "^2.2.0",
+		"date-fns": "^4.1.0",
+		"dayjs": "^1.11.13",
+		"openapi3-ts": "^4.1.2",
+		"zod": "3.23.8",
+		"@typescript-eslint/eslint-plugin": "^6.2.0",
+		"@typescript-eslint/parser": "^6.2.0",
+		"eslint": "8.40.0",
+		"eslint-config-prettier": "^8.8.0",
+		"eslint-import-resolver-typescript": "^3.6.1",
+		"eslint-plugin-import": "^2.29.1",
+		"eslint-plugin-prettier": "^5.0.0",
+		"@types/express": "^4.17.17",
+		"@types/express-requests-logger": "^3.0.4"
+	}
+}
