+++ packages/logger/src/client/.env.schema
