+++ packages/logger/src/client/utils/stringUtils.ts
