+++ packages/logger/src/middleware/reference-log-objects.md
