--- packages/logger/types/package.json
@@ -1,29 +1,0 @@
-{
-	"name": "@pcgl-daco/logger-types",
-	"version": "0.1.0",
-	"description": "",
-	"exports": {
-		"./common": "./dist/common/index.js",
-		"./entities": "./dist/entities/index.js",
-		"./httpResponses": "./dist/httpResponses/index.js"
-	},
-	"type": "module",
-	"engines": {
-		"node": ">=18"
-	},
-	"scripts": {
-		"build": "tsc",
-		"test": "vitest run"
-	},
-	"keywords": [],
-	"author": "",
-	"license": "ISC",
-	"dependencies": {
-		"@anatine/zod-openapi": "^2.2.0",
-		"date-fns": "^4.1.0",
-		"dayjs": "^1.11.13",
-		"@pcgl-daco/logger-eslint-config": "workspace:^",
-		"openapi3-ts": "^4.1.2",
-		"zod": "3.23.8"
-	}
-}
