--- packages/logger/types/tsconfig.json
@@ -1,11 +1,0 @@
-{
-	"extends": "../../../tsconfig.base.json",
-	"compilerOptions": {
-		"outDir": "dist",
-		"rootDir": "src",
-		"baseUrl": ".",
-		"declaration": true
-	},
-	"include": ["src"],
-	"exclude": ["node_modules"]
-}
